+++ sklearn/decomposition/__init__.py
@@ -1,5 +1,7 @@
 """
-Matrix decomposition algorithms
+The :mod:`sklearn.decomposition` module includes matrix decomposition
+algorithms, including among others PCA, NMF or ICA. Most of the algorithms of
+this module can be regarded as dimensionality reduction techniques.
 """
 
 from .nmf import NMF, ProjectedGradientNMF