+++ sklearn/externals/__init__.py
@@ -3,4 +3,3 @@
 External, bundled dependencies.
 
 """
-