+++ sklearn/feature_selection/__init__.py
@@ -1,4 +1,8 @@
-"""Feature slection module for python"""
+"""
+The :mod:`sklearn.feature_selection` module implements feature selection
+algorithms. It currently includes univariate filter selection methods and the
+recursive feature elimination algorithm.
+"""
 
 from .univariate_selection import chi2
 from .univariate_selection import f_classif
@@ -12,4 +16,3 @@ from .univariate_selection import GenericUnivariateSelect
 
 from .rfe import RFE
 from .rfe import RFECV
-