+++ sklearn/linear_model/__init__.py
@@ -1,11 +1,8 @@
 """
-:mod:`sklearn.linear_model` is a module to fit genelarized linear
-models.  It includes Ridge regression, Bayesian Regression, Lasso and
-Elastic Net estimators computed with Least Angle Regression and
-coordinate descent.
-
-It also implements Stochastic Gradient Descent related algorithms.
-
+The :mod:`sklearn.linear_model` module implements genelarized linear models. It
+includes Ridge regression, Bayesian Regression, Lasso and Elastic Net estimators
+computed with Least Angle Regression and coordinate descent. It also implements
+Stochastic Gradient Descent related algorithms.
 """
 
 # See http://scikit-learn.sourceforge.net/modules/sgd.html and
@@ -27,4 +24,3 @@ from .logistic import LogisticRegression
 from .omp import orthogonal_mp, orthogonal_mp_gram, OrthogonalMatchingPursuit
 
 from . import sparse
-