+++ sklearn/svm/classes.py
@@ -27,7 +27,7 @@ class LinearSVC(BaseLibLinear, ClassifierMixin, CoefSelectTransformerMixin):
 
     dual : bool, (default=True)
         Select the algorithm to either solve the dual or primal
-        optimization problem.
+        optimization problem. Prefer dual=False when n_samples > n_features.
 
     tol: float, optional (default=1e-4)
         Tolerance for stopping criteria
@@ -58,7 +58,8 @@ class LinearSVC(BaseLibLinear, ClassifierMixin, CoefSelectTransformerMixin):
 
     Attributes
     ----------
-    `coef_` : array, shape = [n_features] if n_classes == 2 else [n_classes, n_features]
+    `coef_` : array, shape = [n_features] if n_classes == 2 \
+            else [n_classes, n_features]
         Weights asigned to the features (coefficients in the primal
         problem). This is only available in the case of linear kernel.
 
@@ -172,9 +173,9 @@ class SVC(DenseBaseLibSVM, ClassifierMixin):
                  coef0=0.0, shrinking=True, probability=False,
                  tol=1e-3, cache_size=200, scale_C=False):
 
-        DenseBaseLibSVM.__init__(self, 'c_svc', kernel, degree, gamma, coef0,
-                                 tol, C, 0., 0., shrinking, probability,
-                                 cache_size, scale_C)
+        super(SVC, self).__init__('c_svc', kernel, degree, gamma, coef0, tol,
+                                  C, 0., 0., shrinking, probability,
+                                  cache_size, scale_C)
 
 
 class NuSVC(DenseBaseLibSVM, ClassifierMixin):
@@ -277,9 +278,9 @@ class NuSVC(DenseBaseLibSVM, ClassifierMixin):
                  coef0=0.0, shrinking=True, probability=False,
                  tol=1e-3, cache_size=200):
 
-        DenseBaseLibSVM.__init__(self, 'nu_svc', kernel, degree, gamma,
-                                coef0, tol, 0., nu, 0., shrinking, probability,
-                                cache_size, scale_C=None)
+        super(NuSVC, self).__init__('nu_svc', kernel, degree, gamma, coef0,
+                                    tol, 0., nu, 0., shrinking, probability,
+                                    cache_size, scale_C=None)
 
 
 class SVR(DenseBaseLibSVM, RegressorMixin):
@@ -372,9 +373,9 @@ class SVR(DenseBaseLibSVM, RegressorMixin):
                  tol=1e-3, C=1.0, epsilon=0.1, shrinking=True,
                  probability=False, cache_size=200, scale_C=False):
 
-        DenseBaseLibSVM.__init__(self, 'epsilon_svr', kernel, degree, gamma,
-                                 coef0, tol, C, 0., epsilon, shrinking,
-                                 probability, cache_size, scale_C)
+        super(SVR, self).__init__('epsilon_svr', kernel, degree, gamma, coef0,
+                                  tol, C, 0., epsilon, shrinking, probability,
+                                  cache_size, scale_C)
 
     def fit(self, X, y, sample_weight=None, **params):
         """
@@ -397,8 +398,8 @@ class SVR(DenseBaseLibSVM, RegressorMixin):
             Returns self.
         """
         # we copy this method because SVR does not accept class_weight
-        return DenseBaseLibSVM.fit(self, X, y, sample_weight=sample_weight,
-                                  **params)
+        return super(SVR, self).fit(X, y, sample_weight=sample_weight,
+                                    **params)
 
 
 class NuSVR(DenseBaseLibSVM, RegressorMixin):
@@ -410,6 +411,8 @@ class NuSVR(DenseBaseLibSVM, RegressorMixin):
 
     Parameters
     ----------
+    C : float, optional (default=1.0)
+        penalty parameter C of the error term.
 
     nu : float, optional
         An upper bound on the fraction of training errors and a lower bound of
@@ -476,23 +479,23 @@ class NuSVR(DenseBaseLibSVM, RegressorMixin):
     >>> np.random.seed(0)
     >>> y = np.random.randn(n_samples)
     >>> X = np.random.randn(n_samples, n_features)
-    >>> clf = NuSVR(nu=0.1)
+    >>> clf = NuSVR(C=1.0, nu=0.1)
     >>> clf.fit(X, y)
-    NuSVR(cache_size=200, coef0=0.0, degree=3, gamma=0.2, kernel='rbf', nu=0.1,
-       probability=False, shrinking=True, tol=0.001)
+    NuSVR(C=1.0, cache_size=200, coef0=0.0, degree=3, gamma=0.2, kernel='rbf',
+       nu=0.1, probability=False, shrinking=True, tol=0.001)
 
     See also
     --------
     NuSVC, SVR
     """
 
-    def __init__(self, nu=0.5, kernel='rbf', degree=3,
+    def __init__(self, nu=0.5, C=1.0, kernel='rbf', degree=3,
                  gamma=0.0, coef0=0.0, shrinking=True,
                  probability=False, tol=1e-3, cache_size=200):
 
-        DenseBaseLibSVM.__init__(self, 'nu_svr', kernel, degree, gamma, coef0,
-                                 tol, 0, nu, None, shrinking, probability,
-                                 cache_size, scale_C=None)
+        super(NuSVR, self).__init__('nu_svr', kernel, degree, gamma, coef0,
+                                    tol, C, nu, None, shrinking, probability,
+                                    cache_size, scale_C=None)
 
     def fit(self, X, y, sample_weight=None, **params):
         """
@@ -512,7 +515,7 @@ class NuSVR(DenseBaseLibSVM, RegressorMixin):
             Returns self.
         """
         # we copy this method because SVR does not accept class_weight
-        return DenseBaseLibSVM.fit(self, X, y, sample_weight=[], **params)
+        return super(NuSVR, self).fit(X, y, sample_weight=[], **params)
 
 
 class OneClassSVM(DenseBaseLibSVM):
@@ -579,9 +582,10 @@ class OneClassSVM(DenseBaseLibSVM):
     """
     def __init__(self, kernel='rbf', degree=3, gamma=0.0, coef0=0.0,
                  tol=1e-3, nu=0.5, shrinking=True, cache_size=200):
-        DenseBaseLibSVM.__init__(self, 'one_class', kernel, degree, gamma,
-                                 coef0, tol, 0., nu, 0., shrinking, False,
-                                 cache_size, scale_C=None)
+
+        super(OneClassSVM, self).__init__('one_class', kernel, degree, gamma,
+                                          coef0, tol, 0., nu, 0., shrinking,
+                                          False, cache_size, scale_C=None)
 
     def fit(self, X, class_weight={}, sample_weight=None, **params):
         """