+++ sklearn/svm/sparse/classes.py
@@ -33,9 +33,9 @@ class SVC(SparseBaseLibSVM, ClassifierMixin):
                  coef0=0.0, shrinking=True, probability=False,
                  tol=1e-3, cache_size=200, scale_C=False):
 
-        SparseBaseLibSVM.__init__(self, 'c_svc', kernel, degree, gamma, coef0,
-                         tol, C, 0., 0.,
-                         shrinking, probability, cache_size, scale_C)
+        super(SVC, self).__init__('c_svc', kernel, degree, gamma, coef0, tol,
+                                  C, 0., 0., shrinking, probability,
+                                  cache_size, scale_C)
 
 
 class NuSVC(SparseBaseLibSVM, ClassifierMixin):
@@ -67,10 +67,9 @@ class NuSVC(SparseBaseLibSVM, ClassifierMixin):
                  coef0=0.0, shrinking=True, probability=False,
                  tol=1e-3, cache_size=200, scale_C=False):
 
-        SparseBaseLibSVM.__init__(self, 'nu_svc', kernel, degree,
-                         gamma, coef0, tol, 0., nu, 0.,
-                         shrinking, probability, cache_size,
-                         scale_C)
+        super(NuSVC, self).__init__('nu_svc', kernel, degree, gamma, coef0,
+                                    tol, 0., nu, 0., shrinking, probability,
+                                    cache_size, scale_C)
 
 
 class SVR(SparseBaseLibSVM, RegressorMixin):
@@ -95,19 +94,17 @@ class SVR(SparseBaseLibSVM, RegressorMixin):
     >>> clf = SVR(C=1.0, epsilon=0.2)
     >>> clf.fit(X, y)
     SVR(C=1.0, cache_size=200, coef0=0.0, degree=3, epsilon=0.2, gamma=0.2,
-      kernel='rbf', nu=0.5, probability=False, scale_C=False, shrinking=True,
+      kernel='rbf', probability=False, scale_C=False, shrinking=True,
       tol=0.001)
     """
 
     def __init__(self, kernel='rbf', degree=3, gamma=0.0, coef0=0.0,
-                 tol=1e-3, C=1.0, nu=0.5, epsilon=0.1,
-                 shrinking=True, probability=False, cache_size=200,
-                 scale_C=False):
+                 tol=1e-3, C=1.0, epsilon=0.1, shrinking=True,
+                 probability=False, cache_size=200, scale_C=False):
 
-        SparseBaseLibSVM.__init__(self, 'epsilon_svr', kernel,
-                         degree, gamma, coef0, tol, C, nu,
-                         epsilon, shrinking, probability, cache_size,
-                         scale_C)
+        super(SVR, self).__init__('epsilon_svr', kernel, degree, gamma, coef0,
+                                  tol, C, 0., epsilon, shrinking, probability,
+                                  cache_size, scale_C)
 
 
 class NuSVR(SparseBaseLibSVM, RegressorMixin):
@@ -129,20 +126,19 @@ class NuSVR(SparseBaseLibSVM, RegressorMixin):
     >>> np.random.seed(0)
     >>> y = np.random.randn(n_samples)
     >>> X = np.random.randn(n_samples, n_features)
-    >>> clf = NuSVR(nu=0.1)
+    >>> clf = NuSVR(nu=0.1, C=1.0)
     >>> clf.fit(X, y)
-    NuSVR(cache_size=200, coef0=0.0, degree=3, epsilon=0.1, gamma=0.2,
+    NuSVR(C=1.0, cache_size=200, coef0=0.0, degree=3, epsilon=0.1, gamma=0.2,
        kernel='rbf', nu=0.1, probability=False, shrinking=True, tol=0.001)
     """
 
-    def __init__(self, nu=0.5, kernel='rbf', degree=3,
+    def __init__(self, nu=0.5, C=1.0, kernel='rbf', degree=3,
                  gamma=0.0, coef0=0.0, shrinking=True, epsilon=0.1,
                  probability=False, tol=1e-3, cache_size=200):
 
-        SparseBaseLibSVM.__init__(self, 'nu_svr', kernel,
-                         degree, gamma, coef0, tol, 1.0, nu,
-                         epsilon, shrinking, probability, cache_size,
-                         scale_C=None)
+        super(NuSVR, self).__init__('nu_svr', kernel, degree, gamma, coef0,
+                                    tol, C, nu, epsilon, shrinking,
+                                    probability, cache_size, scale_C=None)
 
 
 class OneClassSVM(SparseBaseLibSVM):
@@ -161,9 +157,9 @@ class OneClassSVM(SparseBaseLibSVM):
                  tol=1e-3, nu=0.5, shrinking=True,
                  probability=False, cache_size=200, scale_C=False):
 
-        SparseBaseLibSVM.__init__(self, 'one_class', kernel, degree,
-                         gamma, coef0, tol, 0.0, nu, 0.0,
-                         shrinking, probability, cache_size, scale_C)
+        super(OneClassSVM, self).__init__('one_class', kernel, degree, gamma,
+                                          coef0, tol, 0.0, nu, 0.0, shrinking,
+                                          probability, cache_size, scale_C)
 
     def fit(self, X, class_weight=None, sample_weight=None):
         super(OneClassSVM, self).fit(