+++ sklearn/utils/extmath.py
@@ -26,7 +26,8 @@ def _fast_logdet(A):
     It returns -Inf if det(A) is non positive or is not defined.
     """
     # XXX: Should be implemented as in numpy, using ATLAS
-    # http://projects.scipy.org/numpy/browser/trunk/numpy/linalg/linalg.py#L1559
+    # http://projects.scipy.org/numpy/browser/ \
+    #        trunk/numpy/linalg/linalg.py#L1559
     ld = np.sum(np.log(np.diag(A)))
     a = np.exp(ld / A.shape[0])
     d = np.linalg.det(A / a)
@@ -54,49 +55,6 @@ if hasattr(np.linalg, 'slogdet'):
     fast_logdet = _fast_logdet_numpy
 else:
     fast_logdet = _fast_logdet
-
-try:
-    factorial = math.factorial
-except AttributeError:
-    # math.factorial is only available in Python >= 2.6
-    import operator
-
-    def factorial(n):
-        # don't use reduce operator or 2to3 will fail.
-        # ripped from http://www.joelbdalley.com/page.pl?38
-        # Ensure that n is a Natural number
-        n = abs(int(n))
-        if n < 1:
-            n = 1
-
-        # Store n! in variable x
-        x = 1
-
-        # Compute n!
-        for i in range(1, n + 1):
-            x = i * x
-
-        # Return n!
-        return x
-
-
-try:
-    import itertools
-    combinations = itertools.combinations
-except AttributeError:
-    def combinations(seq, r=None):
-        """Generator returning combinations of items from sequence <seq>
-        taken <r> at a time. Order is not significant. If <r> is not given,
-        the entire sequence is returned.
-        """
-        if r == None:
-            r = len(seq)
-        if r <= 0:
-            yield []
-        else:
-            for i in xrange(len(seq)):
-                for cc in combinations(seq[i + 1:], r - 1):
-                    yield [seq[i]] + cc
 
 
 def density(w, **kwargs):
@@ -167,7 +125,8 @@ def fast_svd(M, k, p=None, q=0, transpose='auto', random_state=0):
     ==========
     Finding structure with randomness: Stochastic algorithms for constructing
     approximate matrix decompositions
-    Halko, et al., 2009 (arXiv:909)
+    Halko, et al., 2009
+    http://arxiv.org/abs/arXiv:0909.4061
 
     A randomized algorithm for the decomposition of matrices
     Per-Gunnar Martinsson, Vladimir Rokhlin and Mark Tygert